requirements.txt
streamlit
langchain
openai
python-docx
PyMuPDF
python-dotenv
paddleocr
pillow
opencv-python-headless
numpy
